requirements.txt
Django==4.2.7
django-ninja>=1.5.0
django-ninja-jwt==5.4.3
psycopg2-binary==2.9.11
python-dotenv==1.0.0
structlog==23.2.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

